style: apply clippy fixes (30+ auto + manual fixes)
- Replace .flatten() with .filter_map(|r| r.ok()) on Lines iterators (prevents infinite loop on read error)
- Derive Default for CpuOffloadMode, ModelType, GpuAllocation
- Use ..Default::default() struct init in build_command_line_for_instance
- Extract InstallPlan type alias for complex return type
- Fix redundant field names, needless borrows, unnecessary casts
- Use next_back() instead of last() on DoubleEndedIterators
- Use rsplit_once() instead of rsplitn().nth(1)
- Use range patterns instead of OR patterns
- Use redundant_closure function pointer
- Allow too_many_arguments for calculate_dense/moe_config (9 args, refactor out of scope)
- Collapse nested ifs in server.rs, deployment_wizard.rs
- All 51 tests pass

+++ src/models/deployment.rs
@@ -177,7 +177,7 @@ pub fn generate_default_profile(
                     gpu_indices: vec![i],
                     port: 8080 + i as u16,
                     memory_utilization: mem_util.max(0.70),
-                    max_num_seqs: Some(if i == 0 { 256 } else { 128 - (i as u32 * 32) }),
+                    max_num_seqs: Some(if i == 0 { 256 } else { 128 - (i * 32) }),
                     model_path: model_path.to_string(),
                     context_size: 4096,
                     provider: provider.to_string(),
